abstract post components, combined title+body+isDraft states into post state, create editPost function
