+++ packages/client/package.json
@@ -7,6 +7,7 @@
     "@testing-library/jest-dom": "^5.14.1",
     "@testing-library/react": "^11.2.7",
     "@testing-library/user-event": "^12.8.3",
+    "axios": "^0.21.4",
     "babel-loader": "8.1.0",
     "phosphor-react": "^1.3.1",
     "react": "^17.0.2",
