+++ packages/client/src/components/Body.js
@@ -1,0 +1,9 @@
+import styled from "styled-components";
+
+const Body = styled.section`
+  padding: 4em;
+  color: black;
+  width: 100%;
+`;
+
+export default Body;
