+++ packages/client/src/components/Button.js
@@ -1,0 +1,16 @@
+import styled from "styled-components";
+
+const Button = styled.button`
+  background-color: black;
+  border: none;
+  color: white;
+  width: 12rem;
+  padding: 1rem 0;
+  &:hover {
+    background-color: #e5e5e5;
+    color: black;
+  }
+  font-size: 1.875rem;
+`;
+
+export default Button;
