+++ packages/client/src/components/ButtonDiv.js
@@ -1,0 +1,9 @@
+import styled from "styled-components";
+
+const ButtonDiv = styled.div`
+  display: flex;
+  justify-content: space-evenly;
+  margin-top: 1.5rem;
+`;
+
+export default ButtonDiv;
