+++ packages/client/src/components/CollapsibleDiv.js
@@ -3,7 +3,13 @@ import styled from "styled-components";
 const CollapsibleDiv = styled.div`
   max-height: ${(props) => (props.hidden ? "0" : "7500px")};
   padding: ${(props) => (props.hidden ? "0" : "1.5rem")};
-  transition: all 1000ms;
+  opacity: ${(props) => (props.hidden ? "0" : "1")};
+  transition: ${(props) =>
+    props.hidden
+      ? "opacity 150ms ease-in-out 150ms"
+      : "opacity 100ms ease-in-out"};
+  transition: ${(props) =>
+    props.hidden ? "max-height 50ms" : "max-height 2000ms"};
   overflow: hidden;
   display: flex;
   flex-flow: column nowrap;
