+++ packages/client/src/components/PostDiv.js
@@ -1,0 +1,12 @@
+import styled from "styled-components";
+
+const PostDiv = styled.div`
+  width: 100%;
+  padding: 1.5rem;
+  display: flex;
+  flex-flow: column nowrap;
+  justify-content: center;
+  align-content: center;
+`;
+
+export default PostDiv;
