+++ packages/client/src/components/PostHeader.js
@@ -1,0 +1,16 @@
+import styled from "styled-components";
+
+const PostHeader = styled.div`
+  text-align: left;
+  background-color: #e5e5e5;
+  padding: 1rem 3rem;
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  z-index: 1000;
+  h2 {
+    font-size: 3rem;
+  }
+`;
+
+export default PostHeader;
