--- packages/client/src/components/mockuser.js
@@ -1,19 +1,0 @@
-
-
-const mockPost = {
-    title: 'First Post',
-    body: "This is a body",
-    isDraft: true,
-    _id: 6456132589,
-    userID: 123    
-}
-
-
-/* 
-userID: {
-    type: ObjectID,
-    Ref: User
-}
-*/
-
-export default mockPost;
