+++ packages/client/src/markdown/Content.mdx
@@ -1,4 +1,4 @@
-import mockPost from "../components/mockuser";
+import mockPost from "../components/mockPost";
 
 export const Box = () => (
   <div style={{ padding: 20, backgroundColor: "limegreen" }} />
